+++ moor_generator/lib/src/model/column.dart
@@ -170,15 +170,6 @@ class MoorColumn implements HasDeclaration {
         ColumnType.real: 'GeneratedRealColumn',
       }[type];
 
-  /// Whether this column is required for insert statements, meaning that a
-  /// non-absent value must be provided for an insert statement to be valid.
-  bool get requiredDuringInsert {
-    final hasDefault = defaultArgument != null || clientDefaultCode != null;
-    final aliasForPk = type == ColumnType.integer &&
-        features.any((f) => f is PrimaryKey || f is AutoIncrement);
-    return !nullable && !hasDefault && !aliasForPk;
-  }
-
   /// The class inside the moor library that represents the same sql type as
   /// this column.
   String get sqlTypeName => sqlTypes[type];