+++ moor/CHANGELOG.md
@@ -14,6 +14,8 @@
 - __Breaking__: Remove the second type variable on `Expression` and subclasses.
 - __Breaking__: Remove `customSelectStream` from `QueryEngine`. The `customSelect`
   method now returns an `Selectable` (like `customSelectQuery`, which in turn has been deprecated).
+- __Breaking__: Columns that are aliases to sqlite's `rowid` column are now longer considered required
+  for inserts
 - Experimentally support IndexedDB to store sqlite data on the web
 - Moor will no longer wait for query stream listeners to receive a done event when closing a database
   or transaction.